refactor: clean the lab
